+++ .phan/config.php
@@ -28,6 +28,11 @@ use Phan\Issue;
  */
 
 return [
+    // The PHP version that the codebase will be checked for compatibility against.
+    // For best results, the PHP binary used to run Phan should have the same PHP version.
+    // (Phan relies on Reflection for some types, param counts,
+    // and checks for undefined classes/methods/functions)
+    //
     // Supported values: `'5.6'`, `'7.0'`, `'7.1'`, `'7.2'`, `'7.3'`, `'7.4'`, `null`.
     // If this is set to `null`,
     // then Phan assumes the PHP version which is closest to the minor version
@@ -36,6 +41,14 @@ return [
     // Note that the **only** effect of choosing `'5.6'` is to infer that functions removed in php 7.0 exist.
     // (See `backward_compatibility_checks` for additional options)
     'target_php_version' => null,
+
+    // Supported values: `'5.6'`, `'7.0'`, `'7.1'`, `'7.2'`, `'7.3'`, `'7.4'`, `null`.
+    // If this is set to `null`, Phan will first attempt to infer the value from
+    // the project's composer.json's `{"require": {"php": "version range"}}` if possible.
+    // If that could not be determined, then Phan assumes `target_php_version`.
+    //
+    // For Phan 3.x, this is determined to be `'7.2'` from `"version": "^7.2.0"`.
+    'minimum_target_php_version' => null,
 
     // Default: true. If this is set to true,
     // and target_php_version is newer than the version used to run Phan,